+++ releases.md
@@ -1,3 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Tidy up request and response body handling.
+
 ## v0.1.0
+
+  - Initial hack.